+++ lib/layers/background_blur_layer.dart
@@ -7,11 +7,13 @@ import 'package:image_editor_plus/data/layer.dart';
 class BackgroundBlurLayer extends StatefulWidget {
   final BackgroundBlurLayerData layerData;
   final VoidCallback? onUpdate;
+  final bool editable;
 
   const BackgroundBlurLayer({
     super.key,
     required this.layerData,
     this.onUpdate,
+    this.editable = false,
   });
 
   @override