+++ lib/layers/emoji_layer.dart
@@ -7,11 +7,13 @@ import 'package:image_editor_plus/modules/emoji_layer_overlay.dart';
 class EmojiLayer extends StatefulWidget {
   final EmojiLayerData layerData;
   final VoidCallback? onUpdate;
+  final bool editable;
 
   const EmojiLayer({
     super.key,
     required this.layerData,
     this.onUpdate,
+    this.editable = false,
   });
 
   @override
@@ -31,41 +33,45 @@ class _EmojiLayerState extends State<EmojiLayer> {
       left: widget.layerData.offset.dx,
       top: widget.layerData.offset.dy,
       child: GestureDetector(
-        onTap: () {
-          showModalBottomSheet(
-            shape: const RoundedRectangleBorder(
-              borderRadius: BorderRadius.only(
-                topRight: Radius.circular(10),
-                topLeft: Radius.circular(10),
-              ),
-            ),
-            context: context,
-            backgroundColor: Colors.transparent,
-            builder: (context) {
-              return EmojiLayerOverlay(
-                index: layers.indexOf(widget.layerData),
-                layer: widget.layerData,
-                onUpdate: () {
-                  if (widget.onUpdate != null) widget.onUpdate!();
-                  setState(() {});
-                },
-              );
-            },
-          );
-        },
-        onScaleUpdate: (detail) {
-          if (detail.pointerCount == 1) {
-            widget.layerData.offset = Offset(
-              widget.layerData.offset.dx + detail.focalPointDelta.dx,
-              widget.layerData.offset.dy + detail.focalPointDelta.dy,
-            );
-          } else if (detail.pointerCount == 2) {
-            widget.layerData.size =
-                initialSize + detail.scale * 5 * (detail.scale > 1 ? 1 : -1);
-          }
+        onTap: widget.editable
+            ? () {
+                showModalBottomSheet(
+                  shape: const RoundedRectangleBorder(
+                    borderRadius: BorderRadius.only(
+                      topRight: Radius.circular(10),
+                      topLeft: Radius.circular(10),
+                    ),
+                  ),
+                  context: context,
+                  backgroundColor: Colors.transparent,
+                  builder: (context) {
+                    return EmojiLayerOverlay(
+                      index: layers.indexOf(widget.layerData),
+                      layer: widget.layerData,
+                      onUpdate: () {
+                        if (widget.onUpdate != null) widget.onUpdate!();
+                        setState(() {});
+                      },
+                    );
+                  },
+                );
+              }
+            : null,
+        onScaleUpdate: widget.editable
+            ? (detail) {
+                if (detail.pointerCount == 1) {
+                  widget.layerData.offset = Offset(
+                    widget.layerData.offset.dx + detail.focalPointDelta.dx,
+                    widget.layerData.offset.dy + detail.focalPointDelta.dy,
+                  );
+                } else if (detail.pointerCount == 2) {
+                  widget.layerData.size = initialSize +
+                      detail.scale * 5 * (detail.scale > 1 ? 1 : -1);
+                }
 
-          setState(() {});
-        },
+                setState(() {});
+              }
+            : null,
         child: Transform.rotate(
           angle: widget.layerData.rotation,
           child: Container(