+++ lib/modules/layers_overlay.dart
@@ -81,7 +81,7 @@ class _ManageLayersOverlayState extends State<ManageLayersOverlay> {
                           ? ClipRRect(
                               borderRadius: BorderRadius.circular(8),
                               child: Image.memory(
-                                layer.image.image,
+                                layer.image.bytes,
                                 fit: BoxFit.cover,
                                 width: 40,
                                 height: 40,
@@ -90,7 +90,7 @@ class _ManageLayersOverlayState extends State<ManageLayersOverlay> {
                               ? ClipRRect(
                                   borderRadius: BorderRadius.circular(8),
                                   child: Image.memory(
-                                    layer.file.image,
+                                    layer.image.bytes,
                                     fit: BoxFit.cover,
                                     width: 40,
                                     height: 40,
@@ -207,41 +207,45 @@ class _ManageLayersOverlayState extends State<ManageLayersOverlay> {
                     width: 64,
                     height: 64,
                     child: Center(
-                      child: layer is TextLayerData || layer is EmojiLayerData
-                          ? Text(
-                              layer is TextLayerData
-                                  ? 'T'
-                                  : (layer as EmojiLayerData).text,
-                              style: const TextStyle(
-                                fontSize: 32,
-                                color: Colors.white,
-                                fontWeight: FontWeight.w100,
-                              ),
-                            )
-                          : (layer is ImageLayerData
-                              ? ClipRRect(
-                                  borderRadius: BorderRadius.circular(8),
-                                  child: Image.memory(
-                                    layer.image.image,
-                                    fit: BoxFit.cover,
-                                    width: 40,
-                                    height: 40,
-                                  ))
-                              : (layer is BackgroundLayerData
+                      child: layer is LinkLayerData
+                          ? const Icon(Icons.link,
+                              size: 32, color: Colors.white)
+                          : (layer is TextLayerData || layer is EmojiLayerData
+                              ? Text(
+                                  layer is TextLayerData
+                                      ? 'T'
+                                      : (layer as EmojiLayerData).text,
+                                  style: const TextStyle(
+                                    fontSize: 32,
+                                    color: Colors.white,
+                                    fontWeight: FontWeight.w100,
+                                  ),
+                                )
+                              : (layer is ImageLayerData
                                   ? ClipRRect(
                                       borderRadius: BorderRadius.circular(8),
                                       child: Image.memory(
-                                        layer.file.image,
+                                        layer.image.bytes,
                                         fit: BoxFit.cover,
                                         width: 40,
                                         height: 40,
                                       ))
-                                  : const Text(
-                                      '',
-                                      style: TextStyle(
-                                        color: Colors.white,
-                                      ),
-                                    ))),
+                                  : (layer is BackgroundLayerData
+                                      ? ClipRRect(
+                                          borderRadius:
+                                              BorderRadius.circular(8),
+                                          child: Image.memory(
+                                            layer.image.bytes,
+                                            fit: BoxFit.cover,
+                                            width: 40,
+                                            height: 40,
+                                          ))
+                                      : const Text(
+                                          '',
+                                          style: TextStyle(
+                                            color: Colors.white,
+                                          ),
+                                        )))),
                     ),
                   ),
                   SizedBox(
@@ -249,7 +253,14 @@ class _ManageLayersOverlayState extends State<ManageLayersOverlay> {
                     child: Column(
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: [
-                        if (layer is TextLayerData)
+                        if (layer is LinkLayerData)
+                          Text(
+                            layer.text,
+                            style: const TextStyle(
+                              color: Colors.white,
+                            ),
+                          )
+                        else if (layer is TextLayerData)
                           Text(
                             layer.text,
                             style: const TextStyle(